chore: start moving shadows into its own feature
